@mxm 2020-05-04
Thanks @lukecwik! This look very promising. I'm going to test this out with one of our applications which experienced problems with the dynamic thread allocation.