@mxm 2020-05-07
The load test can't work properly yet here because it depends on additional code changes in #11558. So good to go.